Allow usage of ratchet/rfc6455 0.4+
ratchet/rfc6455 as a new required param in the constructor of
ServerNegotiator which is breaking the code. As this project still
supports very old php versions we need to keep support for 0.3 but
also include support for 0.4. Reflection is used to check for the
correct version.

The option of using composer's new version checks is not possible as
the older versions of composer do not have this class.